+++ plugit_proxy/views.py
@@ -434,6 +434,11 @@ def handle_special_cases(request, data, baseURI, meta):
     if data.__class__.__name__ == 'PlugIt500':
         return gen500(request, baseURI)
 
+    if data.__class__.__name__ == 'PlugItSpecialCode':
+        r = HttpResponse('')
+        r.status_code = data.code
+        return r
+
     if data.__class__.__name__ == 'PlugItRedirect':
         url = data.url
         if not data.no_prefix:
@@ -805,7 +810,7 @@ def main(request, query, hproPk=None, returnMenuOnly=False):
     current_session = get_current_session(request, hproPk)
 
     # Do the action
-    (data, session_to_set) = plugIt.doAction(query, request.method, getParameters, postParameters, files, things_to_add,
+    (data, session_to_set, headers_to_set) = plugIt.doAction(query, request.method, getParameters, postParameters, files, things_to_add,
                                              proxyMode=proxyMode, session=current_session)
 
     update_session(request, session_to_set, hproPk)
@@ -813,6 +818,10 @@ def main(request, query, hproPk=None, returnMenuOnly=False):
     # Handle special case (redirect, etc..)
     spe_cases = handle_special_cases(request, data, baseURI, meta)
     if spe_cases:
+
+        for header, value in headers_to_set.items():
+            spe_cases[header] = value
+
         return spe_cases
 
     # Save data for proxyMode
@@ -843,7 +852,12 @@ def main(request, query, hproPk=None, returnMenuOnly=False):
         return menu
 
     # Return the final response
-    return build_final_response(request, meta, result, menu, hproject, proxyMode, context)
+    final = build_final_response(request, meta, result, menu, hproject, proxyMode, context)
+
+    for header, value in headers_to_set.items():
+        final[header] = value
+
+    return final
 
 
 def _get_subscription_labels(user, hproject):